First port to parallel compression

+++ webgraph/src/graphs/bvgraph/offset_deg_iter.rs
@@ -193,7 +193,7 @@ impl<D: Decode> OffsetDegIter<D> {
         }
 
         // decode the extra nodes if needed
-        self.decoder.num_of_residuals(nodes_left_to_decode as usize);
+        self.decoder.num_of_residuals(nodes_left_to_decode);
         if nodes_left_to_decode != 0 {
             // pre-allocate with capacity for efficiency
             let _ = self.decoder.read_first_residual();
